Extended multiple screen API and other changes/cleanups
* Android & iOS: Screen hotplug events
* Android & iOS: Create additional windows on the first external screen
* Android: Add Java helpers for Presentation and DisplayListener classes
* Simulate frame time in cases where the OS doesn't provide it
* Use a dummy pbuffer or surfaceless context so OpenGL functions work if no window surfaces exist
* INSTALL doc clarifications
* Improved custom assert() handling

+++ imagine/include/imagine/base/iphone/IOSWindow.hh
@@ -18,6 +18,7 @@
 #include <imagine/engine-globals.h>
 #include <imagine/util/operators.hh>
 #include <imagine/base/iphone/config.h>
+#import <CoreGraphics/CGBase.h>
 
 #ifdef __OBJC__
 #import <imagine/base/iphone/EAGLView.hh>
@@ -39,9 +40,9 @@ public:
 	void *glView_ = nullptr; // EAGLView in ObjC
 	IG::WindowRect contentRect; // active window content
 	#ifdef CONFIG_BASE_IOS_RETINA_SCALE
-	uint pointScale = 1;
+	CGFloat pointScale = 1.;
 	#else
-	static constexpr uint pointScale = 1;
+	static constexpr CGFloat pointScale = 1.;
 	#endif
 
 	constexpr IOSWindow() {}
